Compile server pipeline with -fno-exceptions, split unit tests for client and server
