+++ LibCarla/source/carla/client/ActorBlueprint.cpp
@@ -6,6 +6,7 @@
 
 #include "carla/client/ActorBlueprint.h"
 
+#include "carla/Exception.h"
 #include "carla/StringUtil.h"
 
 #include <algorithm>
