+++ LibCarla/source/test/client/test_image.cpp
